refactor(native): isolate dynamic load cache planning
